Change operation prop names, add inverses
It turns out that if d is the right index, we can actually reverse the
inferences about the effects of weakening and substitution on scope!

Also, the lookup functions are pretty general - they probably belong
in the Fin file, or maybe a more general lib file.
They're probably also in the stdlib, but I can deal with that later.