Beginnings of typechecker proof
Getting all the necessary info to case this thing has taken /a lot/ of
rewriting, and I've had to scrap both cumulativity and impredicative
prop (it's much easier to typecheck when types are unique).

Once this is done it still won't quite be correct, since it won't
respect evaluation, but that can be fixed later.