Evaluation, reorg, lemmas
This is a bit of a monumental commit - there wasn't really a good point
to commit between then and now because of the scale of the reorg.

Unscoped.v has been broken into smaller pieces - now just exports the
core exp datatype, weakening, substitution, and some useful tactics for
using both of those

DeBrujin.v contains all the standard DeBrujin lifting / substitution lemmas
Scoping is where the scoping relation lives now, and lemmas having to do
with it (turns out we don't really use them anywhere else)

Anything having to do with the scoping relation has been moved to
Scoping.v

Evaluation.v contains a small step evaluation relation and a proof of
confluence. Right now the proof lives in Prop, but that shouldn't
matter when this is eventually moved to a theory without Prop.

Progress.v contains a proof of the progression property of arbitrary
terms, which /does/ hold computationally.

TypeChecking.v no longer quite lives up to its name. There's a lot of
work still to be done in here before that can be possible. In
particular, there's a lot of plumbing to deal with to finish the proof
of unique-typing-up-to-evaluation and normalization, the first of which
is necessary for typechecking right now.

I should note that this would all be helped by a theory with real
quotient types, but representing those might also be /difficult/.